+++ db/migrate/20180212102256_create_constructorresults.rb
@@ -3,7 +3,7 @@ class CreateConstructorresults < ActiveRecord::Migration[5.1]
     create_table :constructorresults do |t|
       t.integer :race_id
       t.integer :constructor_id
-      t.decimal :points
+      t.integer :points
 
       t.timestamps
     end
